# planner/core: skip TestEncodePlanPerformance to accelerate CI