@tangenta 2020-12-31
The function signature has been changed to `BenchmarkEncodePlan` in master.